requirements.txt
certifi==2018.11.29
chardet==3.0.4
config==0.4.0
confusable-homoglyphs==3.2.0
dateparser==0.7.1
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.0
django-admin-bootstrapped==2.5.7
django-admin-search==0.1.1
django-autocomplete-light==3.3.2
django-bootstrap3==11.0.0
django-decouple==2.1
django-excel==0.0.10
django-heroku==0.3.1
django-import-export==1.2.0
django-jet==1.0.8
django-material==1.5.0
django-registration==2.4.1
django-suit==0.2.25
django-widget-tweaks==1.4.3
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
jdcal==1.4
jet-bridge==0.1.5
jet-django==0.4.8
lml==0.0.9
odfpy==1.4.0
openpyxl==2.4.9
Pillow==5.4.1
psycopg2==2.7.7
pyexcel==0.5.12
pyexcel-io==0.5.14
pyexcel-webio==0.1.4
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
PyYAML==3.13
regex==2019.2.21
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.18
tablib==0.12.1
texttable==1.6.1
tornado==5.1.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.24.1
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
